patch commands with abstract update update lock/vulnerabilities
