+++ package.json
@@ -34,11 +34,11 @@
     "@typescript-eslint/parser": "^5.4.0",
     "copyfiles": "^2.4.1",
     "enzyme": "^3.11.0",
-    "eslint": "7.32.0",
+    "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-jest": "^24.4.0",
     "eslint-plugin-prettier": "^3.4.1",
-    "eslint-plugin-react": "7.24.0",
+    "eslint-plugin-react": "^7.24.0",
     "husky": "^7.0.4",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.3.1",
@@ -48,12 +48,9 @@
     "jest-prosemirror": "^1.0.10",
     "jest-sonar-reporter": "^2.0.0",
     "lint-staged": "^11.1.2",
-    "prettier": "^2.5.0",
-    "style-loader": "^3.3.0",
-    "stylelint": "^13.13.1",
-    "stylelint-config-standard": "^22.0.0",
-    "stylelint-prettier": "^1.2.0",
-    "terser-webpack-plugin": "^5.3.6",
+    "prettier": "^3.0.0",
+    "stylelint-config-standard": "^36.0.0",
+    "stylelint-prettier": "^5.0.0",
     "ts-jest": "^29.0.7",
     "ts-node": "^10.4.0",
     "typescript": "^4.4.4"
