+++ src/index.ts
@@ -2,9 +2,9 @@ export * from './infoIconDialog';
 export * from './infoIconCommand';
 export * from './InfoIconSubMenu';
 export * from './searchInfoIcon';
-export * from './setInfoIcon';
-export * from  './infoIconView';
-export * from  './infoIconNodeSpec';
+export * from './infoIconPlugin';
+export * from './infoIconView';
+export * from './infoIconNodeSpec';
 export * from './ui/AlertInfo';
 export * from './ui/InfoSubMenuIcon';
 export * from './ui/InfoToolButton';
