+++ src/infoIconPlugin.ts
